refactor (gravsearch): Refactor SPARQL transformation classes.
