+++ webapi/src/main/scala/org/knora/webapi/responders/v2/search/gravsearch/GravsearchUtilV2.scala
@@ -23,9 +23,10 @@ import org.knora.webapi._
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.messages.store.triplestoremessages.{SparqlConstructResponse, SparqlSelectResponse, VariableResultsRow}
 import org.knora.webapi.responders.v2.search._
+import org.knora.webapi.responders.v2.search.gravsearch.prequery.NonTriplestoreSpecificGravsearchToPrequeryGenerator
 import org.knora.webapi.responders.v2.search.gravsearch.types._
 import org.knora.webapi.util.IriConversions._
-import org.knora.webapi.util.{ConstructResponseUtilV2, ErrorHandlingMap, SmartIri, StringFormatter}
+import org.knora.webapi.util.{ConstructResponseUtilV2, ErrorHandlingMap, StringFormatter}
 
 /**
   * Utility methods for [[org.knora.webapi.responders.v2.SearchResponderV2]] (Gravsearch).
@@ -34,332 +35,214 @@ object GravsearchUtilV2 {
 
     protected implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
 
+
     /**
-      * Methods and classes for Sparql transformation.
+      * Represents the IRIs of resources and value objects.
+      *
+      * @param resourceIris    resource IRIs.
+      * @param valueObjectIris value object IRIs.
       */
-    object SparqlTransformation {
+    case class ResourceIrisAndValueObjectIris(resourceIris: Set[IRI], valueObjectIris: Set[IRI])
 
-        /**
-          * Transforms the the Knora explicit graph name to GraphDB explicit graph name.
-          *
-          * @param statement the given statement whose graph name has to be renamed.
-          * @return the statement with the renamed graph, if given.
-          */
-        def transformKnoraExplicitToGraphDBExplicit(statement: StatementPattern): Seq[StatementPattern] = {
-            val transformedPattern = statement.copy(
-                pred = statement.pred match {
-                    case iri: IriRef if iri.iri == OntologyConstants.KnoraBase.MatchesTextIndex.toSmartIri => IriRef(OntologyConstants.Ontotext.LuceneFulltext.toSmartIri) // convert to special Lucene property
-                    case other => other // no conversion needed
-                },
-                namedGraph = statement.namedGraph match {
-                    case Some(IriRef(SmartIri(OntologyConstants.NamedGraphs.KnoraExplicitNamedGraph), _)) => Some(IriRef(OntologyConstants.NamedGraphs.GraphDBExplicitNamedGraph.toSmartIri))
-                    case Some(IriRef(_, _)) => throw AssertionException(s"Named graphs other than ${OntologyConstants.NamedGraphs.KnoraExplicitNamedGraph} cannot occur in non-triplestore-specific generated search query SPARQL")
-                    case None => None
-                }
-            )
+    /**
+      * Traverses value property assertions and returns the IRIs of the value objects and the dependent resources, recursively traversing their value properties as well.
+      * This is method is needed in order to determine if the whole graph pattern is still present in the results after permissions checking handled in [[ConstructResponseUtilV2.splitMainResourcesAndValueRdfData]].
+      * Due to insufficient permissions, some of the resources (both main and dependent resources) and/or values may have been filtered out.
+      *
+      * @param valuePropertyAssertions the assertions to be traversed.
+      * @return a [[ResourceIrisAndValueObjectIris]] representing all resource and value object IRIs that have been found in `valuePropertyAssertions`.
+      */
+    def traverseValuePropertyAssertions(valuePropertyAssertions: Map[IRI, Seq[ConstructResponseUtilV2.ValueRdfData]]): ResourceIrisAndValueObjectIris = {
 
-            Seq(transformedPattern)
-        }
+        // look at the value objects and ignore the property IRIs (we are only interested in value instances)
+        val resAndValObjIris: Seq[ResourceIrisAndValueObjectIris] = valuePropertyAssertions.values.flatten.foldLeft(Seq.empty[ResourceIrisAndValueObjectIris]) {
+            (acc: Seq[ResourceIrisAndValueObjectIris], assertion) =>
 
-        /**
-          * Transforms a non triplestore specific SELECT query to a query for GraphDB.
-          */
-        class GraphDBSelectToSelectTransformer extends SelectToSelectTransformer {
-            def transformStatementInSelect(statementPattern: StatementPattern): Seq[StatementPattern] = Seq(statementPattern)
+                if (assertion.nestedResource.nonEmpty) {
+                    // this is a link value
+                    // recursively traverse the dependent resource's values
 
-            def transformStatementInWhere(statementPattern: StatementPattern, inputOrderBy: Seq[OrderCriterion]): Seq[StatementPattern] = {
-                transformKnoraExplicitToGraphDBExplicit(statementPattern)
-            }
+                    val dependentRes: ConstructResponseUtilV2.ResourceWithValueRdfData = assertion.nestedResource.get
 
-            def transformFilter(filterPattern: FilterPattern): Seq[QueryPattern] = Seq(filterPattern)
+                    // recursively traverse the link value's nested resource and its assertions
+                    val resAndValObjIrisForDependentRes: ResourceIrisAndValueObjectIris = traverseValuePropertyAssertions(dependentRes.valuePropertyAssertions)
 
-        }
-
-        /**
-          * Transforms non triplestore specific SELECT query to a query for a triplestore other than GraphDB (e.g., Fuseki, TODO: to be implemented)
-          */
-        class NoInferenceSelectToSelectTransformer extends SelectToSelectTransformer {
-            def transformStatementInSelect(statementPattern: StatementPattern): Seq[StatementPattern] = Seq(statementPattern)
-
-            def transformStatementInWhere(statementPattern: StatementPattern, inputOrderBy: Seq[OrderCriterion]): Seq[StatementPattern] = {
-                // TODO: if OntologyConstants.NamedGraphs.KnoraExplicitNamedGraph occurs, remove it and use property path syntax to emulate inference.
-                Seq(statementPattern)
-            }
-
-            def transformFilter(filterPattern: FilterPattern): Seq[QueryPattern] = Seq(filterPattern)
-
-        }
-
-        /**
-          * Transforms non-triplestore-specific query patterns to GraphDB-specific ones.
-          */
-        class GraphDBConstructToConstructTransformer extends ConstructToConstructTransformer {
-            def transformStatementInConstruct(statementPattern: StatementPattern): Seq[StatementPattern] = Seq(statementPattern)
-
-            def transformStatementInWhere(statementPattern: StatementPattern, inputOrderBy: Seq[OrderCriterion]): Seq[StatementPattern] = {
-                transformKnoraExplicitToGraphDBExplicit(statementPattern)
-            }
-
-            def transformFilter(filterPattern: FilterPattern): Seq[QueryPattern] = Seq(filterPattern)
-        }
-
-        /**
-          * Transforms non-triplestore-specific query patterns for a triplestore that does not have inference enabled.
-          */
-        class NoInferenceConstructToConstructTransformer extends ConstructToConstructTransformer {
-            def transformStatementInConstruct(statementPattern: StatementPattern): Seq[StatementPattern] = Seq(statementPattern)
-
-            def transformStatementInWhere(statementPattern: StatementPattern, inputOrderBy: Seq[OrderCriterion]): Seq[StatementPattern] = {
-                // TODO: if OntologyConstants.NamedGraphs.KnoraExplicitNamedGraph occurs, remove it and use property path syntax to emulate inference.
-                Seq(statementPattern)
-            }
-
-            def transformFilter(filterPattern: FilterPattern): Seq[QueryPattern] = Seq(filterPattern)
-        }
-
-        /**
-          * Creates a syntactically valid variable base name, based on the given entity.
-          *
-          * @param entity the entity to be used to create a base name for a variable.
-          * @return a base name for a variable.
-          */
-        def escapeEntityForVariable(entity: Entity): String = {
-            val entityStr = entity match {
-                case QueryVariable(varName) => varName
-                case IriRef(iriLiteral, _) => iriLiteral.toString
-                case XsdLiteral(stringLiteral, _) => stringLiteral
-                case _ => throw GravsearchException(s"A unique variable name could not be made for ${entity.toSparql}")
-            }
-
-            entityStr.replaceAll("[:/.#-]", "").replaceAll("\\s", "") // TODO: check if this is complete and if it could lead to collision of variable names
-        }
-
-        /**
-          * Creates a unique variable name from the given entity and the local part of a property IRI.
-          *
-          * @param base        the entity to use to create the variable base name.
-          * @param propertyIri the IRI of the property whose local part will be used to form the unique name.
-          * @return a unique variable.
-          */
-        def createUniqueVariableNameFromEntityAndProperty(base: Entity, propertyIri: IRI): QueryVariable = {
-            val propertyHashIndex = propertyIri.lastIndexOf('#')
-
-            if (propertyHashIndex > 0) {
-                val propertyName = propertyIri.substring(propertyHashIndex + 1)
-                QueryVariable(escapeEntityForVariable(base) + "__" + escapeEntityForVariable(QueryVariable(propertyName)))
-            } else {
-                throw AssertionException(s"Invalid property IRI: $propertyIri")
-            }
-        }
-
-        /**
-          * Represents the IRIs of resources and value objects.
-          *
-          * @param resourceIris    resource IRIs.
-          * @param valueObjectIris value object IRIs.
-          */
-        case class ResourceIrisAndValueObjectIris(resourceIris: Set[IRI], valueObjectIris: Set[IRI])
-
-        /**
-          * Traverses value property assertions and returns the IRIs of the value objects and the dependent resources, recursively traversing their value properties as well.
-          * This is method is needed in order to determine if the whole graph pattern is still present in the results after permissions checking handled in [[ConstructResponseUtilV2.splitMainResourcesAndValueRdfData]].
-          * Due to insufficient permissions, some of the resources (both main and dependent resources) and/or values may have been filtered out.
-          *
-          * @param valuePropertyAssertions the assertions to be traversed.
-          * @return a [[ResourceIrisAndValueObjectIris]] representing all resource and value object IRIs that have been found in `valuePropertyAssertions`.
-          */
-        def traverseValuePropertyAssertions(valuePropertyAssertions: Map[IRI, Seq[ConstructResponseUtilV2.ValueRdfData]]): ResourceIrisAndValueObjectIris = {
-
-            // look at the value objects and ignore the property IRIs (we are only interested in value instances)
-            val resAndValObjIris: Seq[ResourceIrisAndValueObjectIris] = valuePropertyAssertions.values.flatten.foldLeft(Seq.empty[ResourceIrisAndValueObjectIris]) {
-                (acc: Seq[ResourceIrisAndValueObjectIris], assertion) =>
-
-                    if (assertion.nestedResource.nonEmpty) {
-                        // this is a link value
-                        // recursively traverse the dependent resource's values
-
-                        val dependentRes: ConstructResponseUtilV2.ResourceWithValueRdfData = assertion.nestedResource.get
-
-                        // recursively traverse the link value's nested resource and its assertions
-                        val resAndValObjIrisForDependentRes: ResourceIrisAndValueObjectIris = traverseValuePropertyAssertions(dependentRes.valuePropertyAssertions)
-
-                        // get the dependent resource's IRI from the current link value's rdf:object, or rdf:subject in case of an incoming link
-                        val dependentResIri: IRI = if (assertion.isIncomingLink) {
-                            assertion.assertions.getOrElse(OntologyConstants.Rdf.Subject, throw InconsistentTriplestoreDataException(s"expected ${OntologyConstants.Rdf.Subject} for link value ${assertion.valueObjectIri}"))
-                        } else {
-                            assertion.assertions.getOrElse(OntologyConstants.Rdf.Object, throw InconsistentTriplestoreDataException(s"expected ${OntologyConstants.Rdf.Object} for link value ${assertion.valueObjectIri}"))
-                        }
-
-                        // append results from recursion and current value object
-                        ResourceIrisAndValueObjectIris(
-                            resourceIris = resAndValObjIrisForDependentRes.resourceIris + dependentResIri,
-                            valueObjectIris = resAndValObjIrisForDependentRes.valueObjectIris + assertion.valueObjectIri
-                        ) +: acc
+                    // get the dependent resource's IRI from the current link value's rdf:object, or rdf:subject in case of an incoming link
+                    val dependentResIri: IRI = if (assertion.isIncomingLink) {
+                        assertion.assertions.getOrElse(OntologyConstants.Rdf.Subject, throw InconsistentTriplestoreDataException(s"expected ${OntologyConstants.Rdf.Subject} for link value ${assertion.valueObjectIri}"))
                     } else {
-                        // not a link value or no dependent resource given (in order to avoid infinite recursion)
-                        // no dependent resource present
-                        // append results for current value object
-                        ResourceIrisAndValueObjectIris(
-                            resourceIris = Set.empty[IRI],
-                            valueObjectIris = Set(assertion.valueObjectIri)
-                        ) +: acc
+                        assertion.assertions.getOrElse(OntologyConstants.Rdf.Object, throw InconsistentTriplestoreDataException(s"expected ${OntologyConstants.Rdf.Object} for link value ${assertion.valueObjectIri}"))
                     }
-            }
 
-            // convert the collection of `ResourceIrisAndValueObjectIris` into one
-            ResourceIrisAndValueObjectIris(
-                resourceIris = resAndValObjIris.flatMap(_.resourceIris).toSet,
-                valueObjectIris = resAndValObjIris.flatMap(_.valueObjectIris).toSet
-            )
-
+                    // append results from recursion and current value object
+                    ResourceIrisAndValueObjectIris(
+                        resourceIris = resAndValObjIrisForDependentRes.resourceIris + dependentResIri,
+                        valueObjectIris = resAndValObjIrisForDependentRes.valueObjectIris + assertion.valueObjectIri
+                    ) +: acc
+                } else {
+                    // not a link value or no dependent resource given (in order to avoid infinite recursion)
+                    // no dependent resource present
+                    // append results for current value object
+                    ResourceIrisAndValueObjectIris(
+                        resourceIris = Set.empty[IRI],
+                        valueObjectIris = Set(assertion.valueObjectIri)
+                    ) +: acc
+                }
         }
 
-        /**
-          * Checks that the correct schema is used in a statement pattern and that the predicate is allowed in Gravsearch.
-          * If the statement is in the CONSTRUCT clause in the complex schema, non-property variables may refer only to resources or Knora values.
-          *
-          * @param statementPattern     the statement pattern to be checked.
-          * @param querySchema          the API v2 ontology schema used in the query.
-          * @param typeInspectionResult the type inspection result.
-          * @param inConstructClause    `true` if the statement is in the CONSTRUCT clause.
-          */
-        def checkStatement(statementPattern: StatementPattern, querySchema: ApiV2Schema, typeInspectionResult: GravsearchTypeInspectionResult, inConstructClause: Boolean = false): Unit = {
-            // Check each entity in the statement.
-            for (entity <- Seq(statementPattern.subj, statementPattern.pred, statementPattern.obj)) {
-                entity match {
-                    case iriRef: IriRef =>
-                        // The entity is an IRI. If it has a schema, check that it's the query schema.
-                        iriRef.iri.checkApiV2Schema(querySchema, throw GravsearchException(s"${iriRef.toSparql} is not in the correct schema"))
+        // convert the collection of `ResourceIrisAndValueObjectIris` into one
+        ResourceIrisAndValueObjectIris(
+            resourceIris = resAndValObjIris.flatMap(_.resourceIris).toSet,
+            valueObjectIris = resAndValObjIris.flatMap(_.valueObjectIris).toSet
+        )
 
-                        // If we're in the CONSTRUCT clause, don't allow rdf, rdfs, or owl IRIs.
-                        if (inConstructClause && iriRef.iri.toString.contains('#')) {
-                            iriRef.iri.getOntologyFromEntity.toString match {
-                                case OntologyConstants.Rdf.RdfOntologyIri |
-                                     OntologyConstants.Rdfs.RdfsOntologyIri |
-                                     OntologyConstants.Owl.OwlOntologyIri =>
-                                    throw GravsearchException(s"${iriRef.toSparql} is not allowed in a CONSTRUCT clause")
+    }
 
-                                case _ => ()
-                            }
+    /**
+      * Checks that the correct schema is used in a statement pattern and that the predicate is allowed in Gravsearch.
+      * If the statement is in the CONSTRUCT clause in the complex schema, non-property variables may refer only to resources or Knora values.
+      *
+      * @param statementPattern     the statement pattern to be checked.
+      * @param querySchema          the API v2 ontology schema used in the query.
+      * @param typeInspectionResult the type inspection result.
+      * @param inConstructClause    `true` if the statement is in the CONSTRUCT clause.
+      */
+    def checkStatement(statementPattern: StatementPattern, querySchema: ApiV2Schema, typeInspectionResult: GravsearchTypeInspectionResult, inConstructClause: Boolean = false): Unit = {
+        // Check each entity in the statement.
+        for (entity <- Seq(statementPattern.subj, statementPattern.pred, statementPattern.obj)) {
+            entity match {
+                case iriRef: IriRef =>
+                    // The entity is an IRI. If it has a schema, check that it's the query schema.
+                    iriRef.iri.checkApiV2Schema(querySchema, throw GravsearchException(s"${iriRef.toSparql} is not in the correct schema"))
+
+                    // If we're in the CONSTRUCT clause, don't allow rdf, rdfs, or owl IRIs.
+                    if (inConstructClause && iriRef.iri.toString.contains('#')) {
+                        iriRef.iri.getOntologyFromEntity.toString match {
+                            case OntologyConstants.Rdf.RdfOntologyIri |
+                                 OntologyConstants.Rdfs.RdfsOntologyIri |
+                                 OntologyConstants.Owl.OwlOntologyIri =>
+                                throw GravsearchException(s"${iriRef.toSparql} is not allowed in a CONSTRUCT clause")
+
+                            case _ => ()
                         }
+                    }
 
-                    case queryVar: QueryVariable =>
-                        // If the entity is a variable and its type is a Knora IRI, check that the type IRI is in the query schema.
-                        typeInspectionResult.getTypeOfEntity(entity) match {
-                            case Some(typeInfo: GravsearchEntityTypeInfo) =>
-                                typeInfo match {
-                                    case propertyTypeInfo: PropertyTypeInfo =>
-                                        propertyTypeInfo.objectTypeIri.checkApiV2Schema(querySchema, throw GravsearchException(s"${entity.toSparql} is not in the correct schema"))
+                case queryVar: QueryVariable =>
+                    // If the entity is a variable and its type is a Knora IRI, check that the type IRI is in the query schema.
+                    typeInspectionResult.getTypeOfEntity(entity) match {
+                        case Some(typeInfo: GravsearchEntityTypeInfo) =>
+                            typeInfo match {
+                                case propertyTypeInfo: PropertyTypeInfo =>
+                                    propertyTypeInfo.objectTypeIri.checkApiV2Schema(querySchema, throw GravsearchException(s"${entity.toSparql} is not in the correct schema"))
 
-                                    case nonPropertyTypeInfo: NonPropertyTypeInfo =>
-                                        nonPropertyTypeInfo.typeIri.checkApiV2Schema(querySchema, throw GravsearchException(s"${entity.toSparql} is not in the correct schema"))
+                                case nonPropertyTypeInfo: NonPropertyTypeInfo =>
+                                    nonPropertyTypeInfo.typeIri.checkApiV2Schema(querySchema, throw GravsearchException(s"${entity.toSparql} is not in the correct schema"))
 
-                                        // If it's a variable that doesn't represent a property, and we're using the complex schema and the statement
-                                        // is in the CONSTRUCT clause, check that it refers to a resource or value.
-                                        if (inConstructClause && querySchema == ApiV2WithValueObjects) {
-                                            val typeIriStr = nonPropertyTypeInfo.typeIri.toString
+                                    // If it's a variable that doesn't represent a property, and we're using the complex schema and the statement
+                                    // is in the CONSTRUCT clause, check that it refers to a resource or value.
+                                    if (inConstructClause && querySchema == ApiV2WithValueObjects) {
+                                        val typeIriStr = nonPropertyTypeInfo.typeIri.toString
 
-                                            if (!(typeIriStr == OntologyConstants.KnoraApiV2WithValueObjects.Resource || OntologyConstants.KnoraApiV2WithValueObjects.ValueClasses.contains(typeIriStr))) {
-                                                throw GravsearchException(s"${queryVar.toSparql} is not allowed in a CONSTRUCT clause")
-                                            }
+                                        if (!(typeIriStr == OntologyConstants.KnoraApiV2WithValueObjects.Resource || OntologyConstants.KnoraApiV2WithValueObjects.ValueClasses.contains(typeIriStr))) {
+                                            throw GravsearchException(s"${queryVar.toSparql} is not allowed in a CONSTRUCT clause")
                                         }
-                                }
+                                    }
+                            }
 
-                            case None => ()
-                        }
+                        case None => ()
+                    }
 
-                    case xsdLiteral: XsdLiteral =>
-                        val literalOK: Boolean = if (inConstructClause) {
-                            // The only literal allowed in the CONSTRUCT clause is the boolean object of knora-api:isMainResource .
-                            if (xsdLiteral.datatype.toString == OntologyConstants.Xsd.Boolean) {
-                                statementPattern.pred match {
-                                    case iriRef: IriRef =>
-                                        val iriStr = iriRef.iri.toString
-                                        iriStr == OntologyConstants.KnoraApiV2Simple.IsMainResource || iriStr == OntologyConstants.KnoraApiV2WithValueObjects.IsMainResource
+                case xsdLiteral: XsdLiteral =>
+                    val literalOK: Boolean = if (inConstructClause) {
+                        // The only literal allowed in the CONSTRUCT clause is the boolean object of knora-api:isMainResource .
+                        if (xsdLiteral.datatype.toString == OntologyConstants.Xsd.Boolean) {
+                            statementPattern.pred match {
+                                case iriRef: IriRef =>
+                                    val iriStr = iriRef.iri.toString
+                                    iriStr == OntologyConstants.KnoraApiV2Simple.IsMainResource || iriStr == OntologyConstants.KnoraApiV2WithValueObjects.IsMainResource
 
-                                    case _ => false
-                                }
-                            } else {
-                                false
+                                case _ => false
                             }
                         } else {
-                            true
+                            false
                         }
+                    } else {
+                        true
+                    }
 
-                        if (!literalOK) {
-                            throw GravsearchException(s"Statement not allowed in CONSTRUCT clause: ${statementPattern.toSparql.trim}")
-                        }
-
-                    case _ => ()
-                }
-            }
-
-            // Check that the predicate is allowed in a Gravsearch query.
-            statementPattern.pred match {
-                case iriRef: IriRef =>
-                    if (forbiddenPredicates.contains(iriRef.iri.toString)) {
-                        throw GravsearchException(s"Predicate ${iriRef.iri.toSparql} cannot be used in a Gravsearch query")
+                    if (!literalOK) {
+                        throw GravsearchException(s"Statement not allowed in CONSTRUCT clause: ${statementPattern.toSparql.trim}")
                     }
 
                 case _ => ()
             }
         }
 
-        /**
-          * Checks that the correct schema is used in a CONSTRUCT clause, that all the predicates used are allowed in Gravsearch,
-          * and that in the complex schema, non-property variables refer only to resources or Knora values.
-          *
-          * @param constructClause      the CONSTRUCT clause to be checked.
-          * @param typeInspectionResult the type inspection result.
-          */
-        def checkConstructClause(constructClause: ConstructClause, typeInspectionResult: GravsearchTypeInspectionResult): Unit = {
-            for (statementPattern <- constructClause.statements) {
-                checkStatement(
-                    statementPattern = statementPattern,
-                    querySchema = constructClause.querySchema.getOrElse(throw AssertionException(s"ConstructClause has no QuerySchema")),
-                    typeInspectionResult = typeInspectionResult,
-                    inConstructClause = true
-                )
-            }
-        }
+        // Check that the predicate is allowed in a Gravsearch query.
+        statementPattern.pred match {
+            case iriRef: IriRef =>
+                if (forbiddenPredicates.contains(iriRef.iri.toString)) {
+                    throw GravsearchException(s"Predicate ${iriRef.iri.toSparql} cannot be used in a Gravsearch query")
+                }
 
-        // A set of predicates that aren't allowed in Gravsearch.
-        val forbiddenPredicates: Set[IRI] =
-            Set(
-                OntologyConstants.Rdfs.Label,
-                OntologyConstants.KnoraApiV2WithValueObjects.AttachedToUser,
-                OntologyConstants.KnoraApiV2WithValueObjects.AttachedToProject,
-                OntologyConstants.KnoraApiV2WithValueObjects.HasPermissions,
-                OntologyConstants.KnoraApiV2WithValueObjects.CreationDate,
-                OntologyConstants.KnoraApiV2WithValueObjects.LastModificationDate,
-                OntologyConstants.KnoraApiV2WithValueObjects.Result,
-                OntologyConstants.KnoraApiV2WithValueObjects.IsEditable,
-                OntologyConstants.KnoraApiV2WithValueObjects.IsLinkProperty,
-                OntologyConstants.KnoraApiV2WithValueObjects.IsLinkValueProperty,
-                OntologyConstants.KnoraApiV2WithValueObjects.IsInherited,
-                OntologyConstants.KnoraApiV2WithValueObjects.OntologyName,
-                OntologyConstants.KnoraApiV2WithValueObjects.MappingHasName,
-                OntologyConstants.KnoraApiV2WithValueObjects.HasIncomingLink,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartYear,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndYear,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartMonth,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndMonth,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartDay,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndDay,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartEra,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndEra,
-                OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasCalendar,
-                OntologyConstants.KnoraApiV2WithValueObjects.TextValueAsHtml,
-                OntologyConstants.KnoraApiV2WithValueObjects.TextValueAsXml,
-                OntologyConstants.KnoraApiV2WithValueObjects.GeometryValueAsGeometry,
-                OntologyConstants.KnoraApiV2WithValueObjects.LinkValueHasTarget,
-                OntologyConstants.KnoraApiV2WithValueObjects.LinkValueHasTargetIri,
-                OntologyConstants.KnoraApiV2WithValueObjects.ListValueAsListNodeLabel,
-                OntologyConstants.KnoraApiV2WithValueObjects.FileValueAsUrl,
-                OntologyConstants.KnoraApiV2WithValueObjects.FileValueHasFilename,
-                OntologyConstants.KnoraApiV2WithValueObjects.StillImageFileValueHasIIIFBaseUrl
-            )
+            case _ => ()
+        }
     }
+
+    /**
+      * Checks that the correct schema is used in a CONSTRUCT clause, that all the predicates used are allowed in Gravsearch,
+      * and that in the complex schema, non-property variables refer only to resources or Knora values.
+      *
+      * @param constructClause      the CONSTRUCT clause to be checked.
+      * @param typeInspectionResult the type inspection result.
+      */
+    def checkConstructClause(constructClause: ConstructClause, typeInspectionResult: GravsearchTypeInspectionResult): Unit = {
+        for (statementPattern <- constructClause.statements) {
+            checkStatement(
+                statementPattern = statementPattern,
+                querySchema = constructClause.querySchema.getOrElse(throw AssertionException(s"ConstructClause has no QuerySchema")),
+                typeInspectionResult = typeInspectionResult,
+                inConstructClause = true
+            )
+        }
+    }
+
+    // A set of predicates that aren't allowed in Gravsearch.
+    val forbiddenPredicates: Set[IRI] =
+        Set(
+            OntologyConstants.Rdfs.Label,
+            OntologyConstants.KnoraApiV2WithValueObjects.AttachedToUser,
+            OntologyConstants.KnoraApiV2WithValueObjects.AttachedToProject,
+            OntologyConstants.KnoraApiV2WithValueObjects.HasPermissions,
+            OntologyConstants.KnoraApiV2WithValueObjects.CreationDate,
+            OntologyConstants.KnoraApiV2WithValueObjects.LastModificationDate,
+            OntologyConstants.KnoraApiV2WithValueObjects.Result,
+            OntologyConstants.KnoraApiV2WithValueObjects.IsEditable,
+            OntologyConstants.KnoraApiV2WithValueObjects.IsLinkProperty,
+            OntologyConstants.KnoraApiV2WithValueObjects.IsLinkValueProperty,
+            OntologyConstants.KnoraApiV2WithValueObjects.IsInherited,
+            OntologyConstants.KnoraApiV2WithValueObjects.OntologyName,
+            OntologyConstants.KnoraApiV2WithValueObjects.MappingHasName,
+            OntologyConstants.KnoraApiV2WithValueObjects.HasIncomingLink,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartYear,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndYear,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartMonth,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndMonth,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartDay,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndDay,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasStartEra,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasEndEra,
+            OntologyConstants.KnoraApiV2WithValueObjects.DateValueHasCalendar,
+            OntologyConstants.KnoraApiV2WithValueObjects.TextValueAsHtml,
+            OntologyConstants.KnoraApiV2WithValueObjects.TextValueAsXml,
+            OntologyConstants.KnoraApiV2WithValueObjects.GeometryValueAsGeometry,
+            OntologyConstants.KnoraApiV2WithValueObjects.LinkValueHasTarget,
+            OntologyConstants.KnoraApiV2WithValueObjects.LinkValueHasTargetIri,
+            OntologyConstants.KnoraApiV2WithValueObjects.ListValueAsListNodeLabel,
+            OntologyConstants.KnoraApiV2WithValueObjects.FileValueAsUrl,
+            OntologyConstants.KnoraApiV2WithValueObjects.FileValueHasFilename,
+            OntologyConstants.KnoraApiV2WithValueObjects.StillImageFileValueHasIIIFBaseUrl
+        )
 
     object FulltextSearch {
 
@@ -560,8 +443,6 @@ object GravsearchUtilV2 {
           */
         def collectValueVariablesForResource(constructClause: ConstructClause, resource: Entity, typeInspectionResult: GravsearchTypeInspectionResult, variableConcatSuffix: String): Set[QueryVariable] = {
 
-            import SparqlTransformation._
-
             // make sure resource is a query variable or an IRI
             resource match {
                 case queryVar: QueryVariable => ()
@@ -605,7 +486,7 @@ object GravsearchUtilV2 {
                             // linking prop: get value object var and information which values are requested for dependent resource
 
                             // link value object variable
-                            val valObjVar = createUniqueVariableNameFromEntityAndProperty(statementPattern.obj, OntologyConstants.KnoraBase.LinkValue)
+                            val valObjVar = SparqlTransformer.createUniqueVariableNameFromEntityAndProperty(statementPattern.obj, OntologyConstants.KnoraBase.LinkValue)
 
                             // return link value object variable and value objects requested for the dependent resource
                             Set(QueryVariable(valObjVar.variableName + variableConcatSuffix))
@@ -756,7 +637,7 @@ object GravsearchUtilV2 {
           * @return a [[DependentResourcesPerMainResource]].
           */
         def getDependentResourceIrisPerMainResource(prequeryResponse: SparqlSelectResponse,
-                                                    transformer: NonTriplestoreSpecificConstructToSelectTransformer,
+                                                    transformer: NonTriplestoreSpecificGravsearchToPrequeryGenerator,
                                                     mainResourceVar: QueryVariable): DependentResourcesPerMainResource = {
 
             // variables representing dependent resources
@@ -809,7 +690,7 @@ object GravsearchUtilV2 {
           * @return [[ValueObjectVariablesAndValueObjectIris]].
           */
         def getValueObjectVarsAndIrisPerMainResource(prequeryResponse: SparqlSelectResponse,
-                                                     transformer: NonTriplestoreSpecificConstructToSelectTransformer,
+                                                     transformer: NonTriplestoreSpecificGravsearchToPrequeryGenerator,
                                                      mainResourceVar: QueryVariable): ValueObjectVariablesAndValueObjectIris = {
 
             // value objects variables present in the prequery's WHERE clause
@@ -867,8 +748,6 @@ object GravsearchUtilV2 {
                                                     dependentResourceIrisPerMainResource: DependentResourcesPerMainResource,
                                                     valueObjectVarsAndIrisPerMainResource: ValueObjectVariablesAndValueObjectIris,
                                                     requestingUser: UserADM): Map[IRI, ConstructResponseUtilV2.ResourceWithValueRdfData] = {
-
-            import SparqlTransformation._
 
             // separate main resources and value objects (dependent resources are nested)
             // this method removes resources and values the requesting users has insufficient permissions on (missing view permissions).
@@ -943,7 +822,7 @@ object GravsearchUtilV2 {
                                                               valueObjectVarsAndIrisPerMainResource: ValueObjectVariablesAndValueObjectIris,
                                                               allResourceVariablesFromTypeInspection: Set[QueryVariable],
                                                               dependentResourceIrisFromTypeInspection: Set[IRI],
-                                                              transformer: NonTriplestoreSpecificConstructToSelectTransformer,
+                                                              transformer: NonTriplestoreSpecificGravsearchToPrequeryGenerator,
                                                               typeInspectionResult: GravsearchTypeInspectionResult,
                                                               inputQuery: ConstructQuery): Map[IRI, ConstructResponseUtilV2.ResourceWithValueRdfData] = {
